requirements.txt
streamlit
g2pk
google-generativeai